requirements.txt
Django==4.2.7
django-cors-headers==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
Pillow
whitenoise==6.6.0
psycopg==3.1.9
psycopg2-binary==2.9.6
python-dotenv